data_mining/requirements.txt
beautifulsoup4==4.12.3
Scrapy==2.11.1
selenium==4.18.1
pandas==2.2.1
python-dotenv==1.0.1
SQLAlchemy==2.0.28
psycopg2-binary==2.9.9
openai==1.14.1
webdriver-manager==4.0.1
tqdm==4.62.0
bs4==0.0.2
requests==2.31.0
lxml==5.1.0